chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` file. It adds the `firebase-admin` package for integrating Firebase services and the `multer` package for handling file uploads. It also updates the `uuid` package to the latest version. These updates ensure that the project has the latest features and bug fixes from these dependencies.