refactor: Add photo field to createUser function
The code changes in the `auth.js` file add a new field `photo` to the `createUser` function. This field is populated with the user's profile photo URL obtained from the Google login. This addition allows for storing and displaying the user's profile photo in the application.

This commit message follows the convention of starting with a type (`refactor` for code refactoring) followed by a concise and descriptive message of the changes made.